Declare MediatR notifications in Backend project
