+++ HomeRunTracker.Backend/Models/Notifications/ScoringPlayUpdatedNotification.cs
@@ -1,7 +1,8 @@
 ﻿using MediatR;
 
-namespace HomeRunTracker.Core.Actions.ScoringPlays.Notifications;
+namespace HomeRunTracker.Backend.Models.Notifications;
 
+[GenerateSerializer]
 public class ScoringPlayUpdatedNotification : INotification
 {
     public ScoringPlayUpdatedNotification(string homeRunHash, int gameId, DateTimeOffset gameStartTime, string highlightUrl)
@@ -11,12 +12,16 @@ public class ScoringPlayUpdatedNotification : INotification
         GameStartTime = gameStartTime;
         HighlightUrl = highlightUrl;
     }
-    
+
+    [Id(0)]
     public string HomeRunHash { get; }
 
+    [Id(1)]
     public int GameId { get; }
 
+    [Id(2)]
     public DateTimeOffset GameStartTime { get; }
 
+    [Id(3)]
     public string HighlightUrl { get; }
 }
